% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINTER_functions.r
\name{ENCLfunc}
\alias{ENCLfunc}
\title{Summarize single-cell ATAC-seq data according to ENCODE cluster features}
\usage{
ENCLfunc(path, gr_ENCL, type = "paired")
}
\arguments{
\item{path}{Directory to bam files of scATAC-seq data}

\item{gr_ENCL}{GRanges object for ENCODE clusters}

\item{type}{Sequencing read type. "paired" or "single" end read.}
}
\value{
\item{data}{Log2-transformed summarized data matrix.}
 \item{lib}{Total read counts (i.e., library size) for each single cell.}
}
\description{
This function is used for summarizing scATAC-seq data according to ENCODE clusters.
}
\examples{
\dontrun{
scATAC_data <- ENCLfunc("/bam_file_path",gr_ENCL,type="paired")
}
}
\keyword{scATAC-seq}
\keyword{summarize}
