% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINTER_functions.R, R/alignment_test.R
\name{eval_neighbor_test}
\alias{eval_neighbor_test}
\title{Evaluate the spacial discribution of the mixed single cells from different data types while given the cell type labels}
\usage{
eval_neighbor_test(input_data, ref_data, input_mem, ref_mem, dist_scale = 10,
  print_message = TRUE)

eval_neighbor_test(input_data, ref_data, input_mem, ref_mem, dist_scale = 10,
  print_message = TRUE)
}
\arguments{
\item{input_data}{Low dimensional representation of single cell from one data type as the input for matching (e.g., PCs from scRNA-seq data).}

\item{ref_data}{Low dimensional representation of single cell from another data type as the reference for matching (e.g., PCs from scATAC-seq data).}

\item{input_mem}{Cell type label for the input_data.}

\item{ref_mem}{Cell type label for the ref_data.}

\item{dist_scale}{Scale used to define the radius of the region for testing.}

\item{print_message}{Flag to print the radius used for the testing.}

\item{input_data}{Low dimensional representation of single cell from one data type as the input for matching (e.g., PCs from scRNA-seq data).}

\item{ref_data}{Low dimensional representation of single cell from another data type as the reference for matching (e.g., PCs from scATAC-seq data).}

\item{input_mem}{Cell type label for the input_data.}

\item{ref_mem}{Cell type label for the ref_data.}

\item{dist_scale}{Scale used to define the radius of the region for testing.}

\item{print_message}{Flag to print the radius used for the testing.}
}
\value{
\item{test_stat}{P-values from fisher's extract tests.}

\item{test_stat}{P-values from fisher's extract tests.}
}
\description{
This function is used for evaluating whether single cells from the same cell type are mixed well.
For each cell in the input data, a fisher's extract test will be performed to test whether the ratio of input cells to reference cells in the given region is the same as the ratio of the total number of input cells to the total number of reference cells while only single cells from the same cell type are considered.

This function is used for evaluating whether single cells from the same cell type are mixed well.
For each cell in the input data, a fisher's extract test will be performed to test whether the ratio of input cells to reference cells in the given region is the same as the ratio of the total number of input cells to the total number of reference cells while only single cells from the same cell type are considered.
}
\examples{
\dontrun{
test_stat <- eval_neighbor_test(input_data,ref_data,input_mem,ref_mem,dist_scale=10,print_message=TRUE)
}
\dontrun{
test_stat <- eval_neighbor_test(input_data,ref_data,input_mem,ref_mem,dist_scale=10,print_message=TRUE)
}
}
\keyword{evaluation}
\keyword{spacial}
\keyword{test}
