% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_functions.R
\name{mnn_opt}
\alias{mnn_opt}
\title{Search for the optimal parameters for MNN}
\usage{
mnn_opt(data_ref, data_in, k_range = c(16:25), sigma_range = c(0.01, 1),
  dim = 3, dist_scale_in = 10, tol_er = 0.001, ncore = 10,
  fast = FALSE)
}
\arguments{
\item{data_ref}{Single cell data from one data type as the reference for matching (e.g., scATAC-seq data).}

\item{data_in}{Single cell data from another data type as the input for matching (e.g., scRNA-seq data).}

\item{k_range}{Searching space for k, the number of mutual nearest neighbor in MNN.}

\item{sigma_range}{Searching space for sigma, the bandwidth of the Gaussian smoothing kernel used to compute the correction vector.}

\item{dim}{Number of dimension used for matching the single cells. For example, the number of principal components.}

\item{dist_scale_in}{Scale used to define the radius of the region for testing.}

\item{tol_er}{The desired accuracy in function optimize.}

\item{ncore}{Number of CPU cores used for parallel processing. Use ncore = 1 to run the function without parallel processing.}

\item{fast}{A flag indicates whether or not to use a fast neighbor_test.}
}
\value{
\item{k_opt}{The optimal value for k.}
 \item{sigma_opt}{The optimal value for sigma.}
 \item{max_obj}{The average p-value based on the optimal parameters.}
}
\description{
This function is used for searching for the optimal parameters used in MNN.
The goal is find parameters that maximize the average of p-values from the neighbor_test function.
}
\examples{
\dontrun{
result_opt <- mnn_opt(data_ref,data_in,k_range=c(16:25),sigma_range=c(0.01,1),dim=3,dist_scale_in=10,tol_er=0.001,ncore=10)
}
}
\keyword{MNN}
\keyword{optimization}
