% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINTER_functions.R, R/alignment_test.R
\name{neighbor_test}
\alias{neighbor_test}
\title{Evaluate the spacial discribution of the mixed single cells from different data types}
\usage{
neighbor_test(input_data, ref_data, dist_scale = 10,
  print_message = TRUE)

neighbor_test(input_data, ref_data, dist_scale = 10,
  print_message = TRUE)
}
\arguments{
\item{input_data}{Low dimensional representation of single cell from one data type as the input for matching (e.g., PCs from scRNA-seq data).}

\item{ref_data}{Low dimensional representation of single cell from another data type as the reference for matching (e.g., PCs from scATAC-seq data).}

\item{dist_scale}{Scale used to define the radius of the region for testing.}

\item{print_message}{Flag to print the radius used for the testing.}

\item{input_data}{Low dimensional representation of single cell from one data type as the input for matching (e.g., PCs from scRNA-seq data).}

\item{ref_data}{Low dimensional representation of single cell from another data type as the reference for matching (e.g., PCs from scATAC-seq data).}

\item{dist_scale}{Scale used to define the radius of the region for testing.}

\item{print_message}{Flag to print the radius used for the testing.}
}
\value{
\item{input_count}{The number of cells in the input data for each test.}
 \item{ref_count}{The number of cells in the reference data for each test.}
 \item{pval}{P-values from fisher's extract tests.}

\item{input_count}{The number of cells in the input data for each test.}
 \item{ref_count}{The number of cells in the reference data for each test.}
 \item{pval}{P-values from fisher's extract tests.}
}
\description{
This function is used for testing whether the single cells from different data types are mixed well.
For each cell in the input data, a fisher's extract test will be performed to test whether the ratio of input cells to reference cells in the given region is the same as the ratio of the total number of input cells to the total number of reference cells.

This function is used for testing whether the single cells from different data types are mixed well.
For each cell in the input data, a fisher's extract test will be performed to test whether the ratio of input cells to reference cells in the given region is the same as the ratio of the total number of input cells to the total number of reference cells.
}
\examples{
\dontrun{
neighbor_test_p <- neighbor_test(input_data,ref_data,dist_scale=10)$pval
}
\dontrun{
neighbor_test_p <- neighbor_test(input_data,ref_data,dist_scale=10)$pval
}
}
\keyword{spacial}
\keyword{test}
