% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINTER_functions.r
\name{pre_model}
\alias{pre_model}
\title{Prediction of ENCODE cluster features based on scRNA-seq data}
\usage{
pre_model(expr_select, DNase_train, RNA_train, num_predictor = 10,
  cluster_scale = 10, seed = 12345)
}
\arguments{
\item{expr_select}{Input gene expression data from scRNA-seq.}

\item{DNase_train}{ENCODE cluster features from DNase-seq data for building the regression model.}

\item{RNA_train}{Gene expression from ENCODE RNA-seq data for building the regression model.}

\item{num_predictor}{Number of predictors used in the prediction model.}

\item{cluster_scale}{The scale to determine the number of gene clusters. The number of gene clusters is obtained by [the number of genes]/[cluster_scale].}

\item{seed}{Set the seed in kmeans clustering for reproducible results.}
}
\value{
\item{Y_pre}{A matrix of predicted ENCODE cluster features.}
}
\description{
This function is used for predicting ENCODE cluster features based on scRNA-seq data. The scRNA-seq data are first clustered into gene clusters and the cluster means are used as predictors.
}
\examples{
\dontrun{
Y_pre <- pre_model(expr_select,DNase_train,RNA_train,num_predictor=10,cluster_scale=10,seed=12345)
}
}
\keyword{prediction}
