% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_functions.R
\name{predict_opt}
\alias{predict_opt}
\title{Search for the optimal parameters for SINTER}
\usage{
predict_opt(atac_data, expr_data, DNase_train, RNA_train,
  num_predictor = c(25, 25, 30), cluster_scale = c(10, 20, 50),
  k_range = c(20:29), sigma_range = c(0.01, 1), k_in = 20,
  sigma_in = 0.1, dim = 3, dist_scale_in = 10, subsample = FALSE,
  MNN_opt = TRUE, fast = FALSE, MNN_ref = "scATAC", tol_er = 0.001,
  ncore = 10, seed = 12345)
}
\arguments{
\item{atac_data}{scATAC-seq data for matching.}

\item{expr_data}{scRNA-seq data for matching.}

\item{DNase_train}{ENCODE cluster features from DNase-seq data for building the regression model.}

\item{RNA_train}{Gene expression from ENCODE RNA-seq data for building the regression model.}

\item{num_predictor}{Searching space for number of predictors used in the regression model.}

\item{cluster_scale}{Searching space for the scale to determine the number of gene clusters.}

\item{k_range}{Searching space for k, the number of mutual nearest neighbor in MNN if flag MNN_opt==TRUE.}

\item{sigma_range}{Searching space for sigma, the bandwidth of the Gaussian smoothing kernel used to compute the correction vector if flag MNN_opt==TRUE.}

\item{k_in}{Setting K, the number of mutual nearest neighbor in MNN if flag MNN_opt!=TRUE.}

\item{sigma_in}{Setting sigma, the bandwidth of the Gaussian smoothing kernel used to compute the correction vector if flag MNN_opt!=TRUE.}

\item{dim}{Number of dimension used for matching the single cells. For example, the number of principal components.}

\item{dist_scale_in}{Scale used to define the radius of the region for testing.}

\item{subsample}{A percentage value to determine whether the paramter searching should be done in a subset of cells instead of using all cells. Set subsample=FALSE to use all cells.}

\item{MNN_opt}{A flag to determine whether the parameters search should be performed for MNN.}

\item{fast}{A flag indicates whether or not to use a fast neighbor_test.}

\item{MNN_ref}{A flag to determine which data type is used as reference in MNN. Select from "scATAC" and "scRNA".}

\item{tol_er}{The desired accuracy in function optimize.}

\item{ncore}{Number of CPU cores used for parallel processing. Use ncore = 1 to run the function without parallel processing.}

\item{seed}{The seed used for subsampling if subsample!=FALSE.}
}
\value{
\item{num_predictor_opt}{The optimal value for number of predictors.}
 \item{cluster_scale_opt}{The optimal value for cluster scale.}
 \item{k_opt}{The optimal value for k.}
 \item{sigma_opt}{The optimal value for sigma.}
 \item{max_obj}{The average p-value based on the optimal parameters.}
}
\description{
This function is used for searching for the optimal parameters used in SINTER.
The goal is find parameters that maximize the average of p-values from the neighbor_test function.
}
\examples{
\dontrun{
result_opt <- predict_opt(atac_data,expr_data,DNase_train,RNA_train,num_predictor=c(25,25,30),cluster_scale=c(10,20,50),k_range=c(20:29),sigma_range=c(0.01,1),
k_in=20,sigma_in=0.1,dim=3,dist_scale_in=10,subsample=FALSE,MNN_opt=TRUE,MNN_ref="scATAC",tol_er=0.001,ncore=10,seed=12345)
}
}
\keyword{MNN}
\keyword{optimization}
