% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINTER_functions.R
\name{run_MNN}
\alias{run_MNN}
\title{Run MNN to correct for platform effects}
\usage{
run_MNN(atac_data, pre_result, k, sigma, MNN_ref = "scATAC")
}
\arguments{
\item{atac_data}{scATAC-seq data for matching.}

\item{pre_result}{Predicted scATAC-seq data based on scRNA-seq.}

\item{k}{The number of mutual nearest neighbor in MNN.}

\item{sigma}{The bandwidth of the Gaussian smoothing kernel used to compute the correction vector.}

\item{MNN_ref}{A flag to determine which data type is used as reference in MNN. Select from "scATAC" and "scRNA".}
}
\value{
\item{data_combine}{The combined data matrix from all single cells.}
}
\description{
This function is used run MNN to correct for platform effects between the experimental and predicted scATAC-seq data.
}
\examples{
\dontrun{
data_combine <- run_MNN(atac_data,pre_result,k=param_opt$k_opt,sigma=param_opt$sigma_opt,MNN_ref="scATAC")
}
}
\keyword{MNN}
