% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINTER_functions.r
\name{select_DHS_cluster}
\alias{select_DHS_cluster}
\title{Select variable ENCODE cluster features from scATAC-seq data}
\usage{
select_DHS_cluster(DHS_data, lib_th = 1000, var_th = 0.2,
  dataplot = FALSE)
}
\arguments{
\item{DHS_data}{Summarized ENCODE cluster features from scATAC-seq data.}

\item{lib_th}{Threshold for filtering out cells with low total number of reads.}

\item{var_th}{Threshold for filtering out ENCODE cluster features with low variability in the given percentage of features.}

\item{dataplot}{Plot the mean and variance of all ENCODE cluster features and the selected features are highlighted.}
}
\value{
\item{data_select}{Selected ENCODE cluster features.}
 \item{select_idx}{Index for the selected ENCODE cluster features.}
}
\description{
This function is used for selecting variable ENCODE cluster features from scATAC-seq data.
}
\examples{
\dontrun{
scATAC_data_select <- select_DHS_cluster(DHS_data,lib_th=1e3,var_th=0.2,dataplot=FALSE)
}
}
\keyword{feature}
\keyword{scATAC-seq}
\keyword{variable}
